Actually, fuck it
In production, we can just pull everything down at startup. It's going
to make the startup time monstrous but then the queries will be super
nippy. If you look at search.py, the ranking used is tuned to not pull
everything down, but we can change that now, which is neato.