All the data (#3)
* Phew, finally have tables for everything

* Trying out some new stuff, including scans

* Should be good to go

* Whoops! Accidentally changed a column's name

* Add a reverse index for tags so we can do Jaccard

* Jaccard baby! (hopefully)

* Hopefully fix a circular dependency

* Problems with getting games

* Heroku circular dependency fix

* Default to sensible values for Dynamo configs

* Huge cleanup on repeated code

* Jaccard to be kidding me

* Actually, fuck it

In production, we can just pull everything down at startup. It's going
to make the startup time monstrous but then the queries will be super
nippy. If you look at search.py, the ranking used is tuned to not pull
everything down, but we can change that now, which is neato.

* Everything is going Jaccarding to plan

* Little bit of magic to handle cold cache

* Normalize name

* Add Steam URL