So this actually works?? We'll have to see
I tested the ranking output on a couple of games, it actually works
rather well. I'm pretty pleased. Now we need to do the query versioning
and actually using the compressed matrix to do the ranking! I think we
should also focus on finding a way to interpret the lower dimensions,
and how we can use those to provide some feedback saying "this game was
picked because of its strengths in this dimension, yadda yadda"